Fixed crash while saving to disk
Calling #shutdown only presists data to the disk if the redis config
contains at least one SAVEPOINT.

When we were shutting the redis server down redis-rb raised a
TimeoutException because the saving was taking too much time. We wrapped
the call to shutdown around a begin/rescue block where we loop until we
see confirmation that the Redis process is infact shutdown.